feat: add Redis

+++ src/main/java/com/swp391/skincare_products_sales_system/repository/OriginRepository.java
@@ -3,6 +3,6 @@ package com.swp391.skincare_products_sales_system.repository;
 import com.swp391.skincare_products_sales_system.model.Origin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OriginRepository extends JpaRepository<Origin, Integer> {
+public interface OriginRepository extends JpaRepository<Origin, Long> {
 
 }
